test_infra: Move `struct Guest` and `struct GuestCommand` from tests
In this way, these structs can be reused for performance tests.
