+++ test_infra/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 dirs = "3.0.1"
 epoll = "4.3.1"
+lazy_static= "1.4.0"
 libc = "0.2.91"
 ssh2 = "0.9.1"
 vmm-sys-util = "0.9.0"
